+++ Examples/BasicGame/include/BasicGame.hpp
@@ -2,11 +2,8 @@
 
 #include <Pyramid/Core/Game.hpp>
 #include <Pyramid/Graphics/Camera.hpp>
-#include <Pyramid/Graphics/Geometry/MeshCache.hpp>
-#include <Pyramid/Graphics/Shader/ShaderCache.hpp>
-#include <Pyramid/Graphics/Texture/TextureCache.hpp>
 #include <Pyramid/Graphics/Material/Material.hpp>
-#include <Pyramid/Graphics/Material/MaterialCache.hpp>
+#include <Pyramid/Graphics/Resources/ResourceRegistry.hpp>
 #include <Pyramid/Graphics/Renderer/RenderSystem.hpp>
 #include <Pyramid/Graphics/Scene.hpp>
 
@@ -15,13 +12,9 @@
 namespace Pyramid
 {
     class Mesh;
-    class MeshCache;
-    class ShaderCache;
     class ShaderProgram;
-    class TextureCache;
     class TextureResource;
     class Material;
-    class MaterialCache;
 }
 
 class BasicGame final : public Pyramid::Game
@@ -40,10 +33,6 @@ private:
     bool SetupScene();
     void UpdateCamera(float deltaTime);
 
-    std::unique_ptr<Pyramid::MeshCache> m_meshCache;
-    std::unique_ptr<Pyramid::ShaderCache> m_shaderCache;
-    std::unique_ptr<Pyramid::TextureCache> m_textureCache;
-    std::unique_ptr<Pyramid::MaterialCache> m_materialCache;
     std::unique_ptr<Pyramid::Renderer::RenderSystem> m_renderSystem;
     std::shared_ptr<Pyramid::Scene> m_scene;
     std::unique_ptr<Pyramid::Camera> m_camera;